requirements.txt
intersystems-irispython>=5.0.0
mcp>=0.4.0
pydantic>=2.0.0
fastapi>=0.95.0
fastmcp>=0.9.0
uvicorn[standard]>=0.20.0
